#1776 - Advanced workload and workflow
- Renamed stuff
- Move settings from the dynamic module into the base workload module
- Do not use internal IDs in the workflow selector dropdown
- Removed redundant WorkloadRegistry
