+++ inception-external-search-elastic/src/main/java/de/tudarmstadt/ukp/inception/externalsearch/elastic/ElasticSearchProvider.java
@@ -87,7 +87,7 @@ public class ElasticSearchProvider
             
             QueryBuilder qb = QueryBuilders.simpleQueryStringQuery(aQuery)
                     .field(aTraits.getDefaultField());
-
+            
             if (aTraits.isRandomOrder()) {
                 RandomScoreFunctionBuilder randomFunc = ScoreFunctionBuilders.randomFunction();
                 randomFunc.seed(aTraits.getSeed());
@@ -100,7 +100,6 @@ public class ElasticSearchProvider
             
             SearchRequest searchRequest = new SearchRequest(aTraits.getIndexName())
                     .source(searchSourceBuilder);
-
             SearchResponse response = client.search(searchRequest);
     
             for (SearchHit hit: response.getHits().getHits()) {
