+++ inception-ui-core/pom.xml
@@ -157,14 +157,6 @@
      <artifactId>javax.servlet-api</artifactId>
      <scope>provided</scope> 
     </dependency>
-    <dependency>
-      <groupId>de.tudarmstadt.ukp.inception.app</groupId>
-      <artifactId>inception-workload</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>de.tudarmstadt.ukp.inception.app</groupId>
-      <artifactId>inception-workload-static</artifactId>
-    </dependency>
   </dependencies>
   
   <build>
