+++ inception-workload-dynamic/pom.xml
@@ -25,7 +25,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <artifactId>inception-workload-dynamic</artifactId>
-    <name>INCEpTION - Workload-dynamic</name>
+    <name>INCEpTION - Workload Management - Dynamic</name>
     <dependencies>
       <dependency>
         <groupId>de.tudarmstadt.ukp.clarin.webanno</groupId>
