+++ inception-workload-dynamic/src/main/java/de/tudarmstadt/ukp/inception/workload/dynamic/DynamicWorkloadExtension.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2020
+ * Ubiquitous Knowledge Processing (UKP) Lab
+ * Technische Universität Darmstadt
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.tudarmstadt.ukp.inception.workload.dynamic;
+
+import de.tudarmstadt.ukp.inception.workload.dynamic.config.DynamicWorkloadManagerAutoConfiguration;
+import de.tudarmstadt.ukp.inception.workload.extension.WorkloadManagerExtension;
+
+/**
+ * <p>
+ * This class is exposed as a Spring Component via
+ * {@link DynamicWorkloadManagerAutoConfiguration#dynamicWorkloadExtension()}
+ * </p>
+ */
+public class DynamicWorkloadExtension implements WorkloadManagerExtension
+{
+    public static final String DYNAMIC_EXTENSION_ID = "dynamic";
+
+    @Override
+    public String getId()
+    {
+        return DYNAMIC_EXTENSION_ID;
+    }
+
+    @Override
+    public String getLabel()
+    {
+        return "Dynamic assignment";
+    }
+}
