--- inception-workload-static/src/main/java/de/tudarmstadt/ukp/inception/workload/monitoring/config/StaticWorkloadManagerAutoConfiguration.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2020
- * Ubiquitous Knowledge Processing (UKP) Lab
- * Technische Universität Darmstadt
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.tudarmstadt.ukp.inception.workload.monitoring.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.annotation.Order;
-
-import de.tudarmstadt.ukp.inception.workload.monitoring.extension.StaticWorkloadExtension;
-
-@Configuration
-@Order(500)
-public class StaticWorkloadManagerAutoConfiguration
-{
-    @Bean
-    public StaticWorkloadExtension staticWorkloadExtension()
-    {
-        return new StaticWorkloadExtension();
-    }
-}
