--- inception-workload-static/src/main/java/de/tudarmstadt/ukp/inception/workload/monitoring/extension/StaticWorkloadExtension.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2020
- * Ubiquitous Knowledge Processing (UKP) Lab
- * Technische Universität Darmstadt
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.tudarmstadt.ukp.inception.workload.monitoring.extension;
-
-import de.tudarmstadt.ukp.inception.workload.extension.WorkloadExtension;
-import de.tudarmstadt.ukp.inception.workload.monitoring.config.StaticWorkloadManagerAutoConfiguration;
-
-/**
- * <p>
- * This class is exposed as a Spring Component via
- * {@link StaticWorkloadManagerAutoConfiguration#staticWorkloadExtension}
- * </p>
- */
-public class StaticWorkloadExtension implements WorkloadExtension
-{
-    public static final String STATIC_EXTENSION_ID = "Static Workload";
-
-    @Override
-    public String getId()
-    {
-        return STATIC_EXTENSION_ID;
-    }
-
-}
